all: create geth namespace with supply delta
